moved fs references to store module
